requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pandas==2.2.3
gunicorn==23.0.0
Werkzeug==3.0.4
